Renames generated to js_interop
