+++ packages/google_sign_in/google_sign_in_web/CHANGELOG.md
@@ -3,6 +3,10 @@
 * Updates code for `no_leading_underscores_for_local_identifiers` lint.
 * Updates minimum Flutter version to 2.10.
 
+## 0.10.2+1
+
+* Renames generated folder to js_interop.
+
 ## 0.10.2
 
 * Migrates to new platform-interface `initWithParams` method.
